wip: can run eclipse on flow_example
