--- src/ert/resources/forward_models/res/script/ecl100_config.yml
@@ -1,14 +1,0 @@
-# NB: There are inter related dependencies between this file, the EclConfig
-#     class which essentially internalizes this file and the EclRun class
-#     which uses the EclConfig class.
-#
-# You can have a shared 'env' attribute which is an environment variable map
-# which will be set before the simulator starts. If there are env settings in
-# the simulator as well they will be merged with these common settings, the
-# simulator specific take presedence.
-
-eclrun_env:
-  SLBSLS_LICENSE_FILE: something@yourcompany.com
-  ECLPATH: /path/to/ecl
-  PATH: /path/to/ecl/macros
-  LSB_JOBID: null
