--- src/ert/resources/forward_models/res/script/ecl_run.py
@@ -1,545 +1,0 @@
-import datetime
-import glob
-import os
-import os.path
-import re
-import shutil
-import socket
-import subprocess
-import sys
-import time
-from argparse import ArgumentParser
-from collections import namedtuple
-from contextlib import contextmanager
-from pathlib import Path
-from random import random
-from typing import List
-
-import resfo
-from ecl_config import EclConfig
-
-from ert.plugins import plugin_manager
-
-
-def ecl_output_has_license_error(ecl_output: str):
-    return (
-        "LICENSE ERROR" in ecl_output
-        or "LICENSE FAILURE" in ecl_output
-        or "not allowed in license" in ecl_output
-    )
-
-
-class EclError(RuntimeError):
-    def failed_due_to_license_problems(self) -> bool:
-        # self.args[0] contains the multiline ERROR messages and SLAVE startup messages
-        if ecl_output_has_license_error(self.args[0]):
-            return True
-        if re.search(a_slave_failed_pattern, self.args[0]):
-            for match in re.finditer(slave_run_paths, self.args[0], re.MULTILINE):
-                (ecl_case_starts_with, ecl_case_dir) = match.groups()
-                for prt_file in glob.glob(
-                    f"{ecl_case_dir}/{ecl_case_starts_with}*.PRT"
-                ):
-                    if ecl_output_has_license_error(
-                        Path(prt_file).read_text(encoding="utf-8")
-                    ):
-                        return True
-        return False
-
-
-EclipseResult = namedtuple("EclipseResult", "errors bugs")
-body_sub_pattern = r"(\s^\s@.+$)*"
-date_sub_pattern = r"\s+AT TIME\s+(?P<Days>\d+\.\d+)\s+DAYS\s+\((?P<Date>(.+)):\s*$"
-error_pattern_e100 = rf"^\s@--  ERROR{date_sub_pattern}${body_sub_pattern}"
-error_pattern_e300 = rf"^\s@--Error${body_sub_pattern}"
-slave_started_pattern = (
-    rf"^\s@--MESSAGE{date_sub_pattern}\s^\s@\s+STARTING SLAVE.+${body_sub_pattern}"
-)
-a_slave_failed_pattern = r"\s@\s+SLAVE RUN.*HAS STOPPED WITH AN ERROR CONDITION.\s*"
-slave_run_paths = r"^\s@\s+STARTING SLAVE\s+[^ ]+RUNNING \([^ ]\)\s*$"
-slave_run_paths = r"\s@\s+STARTING SLAVE .* RUNNING (\w+)\s*^\s@\s+ON HOST.*IN DIRECTORY\s*^\s@\s+(.*)"
-
-
-def make_LSB_MCPU_machine_list(LSB_MCPU_HOSTS):
-    host_numcpu_list = LSB_MCPU_HOSTS.split()
-    host_list = []
-    for index in range(len(host_numcpu_list) // 2):
-        machine = host_numcpu_list[2 * index]
-        host_numcpu = int(host_numcpu_list[2 * index + 1])
-        for _ in range(host_numcpu):
-            host_list.append(machine)
-    return host_list
-
-
-def _expand_SLURM_range(rs):
-    if "-" in rs:
-        tmp = rs.split("-")
-        return range(int(tmp[0]), int(tmp[1]) + 1)
-    else:
-        return [int(rs)]
-
-
-def _expand_SLURM_node(node_string):
-    match_object = re.match(r"(?P<base>[^[]+)\[(?P<range>[-0-9,]+)\]", node_string)
-    if match_object:
-        node_list = []
-        base = match_object.groupdict()["base"]
-        range_string = match_object.groupdict()["range"]
-        for rs in range_string.split(","):
-            for num in _expand_SLURM_range(rs):
-                node_list.append(f"{base}{num}")
-        return node_list
-    else:
-        return [node_string]
-
-
-def _expand_SLURM_task_count(task_count_string):
-    match_object = re.match(r"(?P<count>\d+)(\(x(?P<mult>\d+)\))?", task_count_string)
-    if match_object:
-        match_dict = match_object.groupdict()
-        print(match_dict)
-        count = int(match_dict["count"])
-        mult_string = match_dict["mult"]
-        mult = 1 if mult_string is None else int(mult_string)
-
-        return [count] * mult
-    else:
-        raise ValueError(f"Failed to parse SLURM_TASKS_PER_NODE: {task_count_string}")
-
-
-# The list of available machines/nodes and how many tasks each node should get
-# is available in the slurm environment variables SLURM_JOB_NODELIST and
-# SLURM_TASKS_PER_NODE. These string variables are in an incredibly compact
-# notation, and there are some hoops to expand them. The short description is:
-#
-#  1. They represent flat lists of hostnames and the number of cpu's on that
-#     host respectively.
-#
-#  2. The outer structure is a ',' separated lis.
-#
-#  3. The items in SLURM_JOB_NODELIST have a compact notation
-#     base-[n1-n2,n3-n4] which is expanded to the nodelist: [base-n1,
-#     base-n1+1, base-n1+2, ... , base-n4-1, base-n4]
-#
-#  4. The SLURM_TASK_PER_NODE items has the compact notation 3(x4) which
-#     implies that four consecutive nodes (from the expanded
-#     SLURM_JOB_NODELIST) should have three CPUs each.
-#
-# For further details see the sbatch manual page.
-
-
-def make_SLURM_machine_list(SLURM_JOB_NODELIST, SLURM_TASKS_PER_NODE):
-    # We split on ',' - but not on ',' which is inside a [...]
-    split_re = ",(?![^[]*\\])"
-    nodelist = []
-    for node_string in re.split(split_re, SLURM_JOB_NODELIST):
-        nodelist += _expand_SLURM_node(node_string)
-
-    task_count_list = []
-    for task_count_string in SLURM_TASKS_PER_NODE.split(","):
-        task_count_list += _expand_SLURM_task_count(task_count_string)
-
-    host_list = []
-    for node, count in zip(nodelist, task_count_list):
-        host_list += [node] * count
-
-    return host_list
-
-
-def make_LSB_machine_list(LSB_HOSTS):
-    return LSB_HOSTS.split()
-
-
-@contextmanager
-def pushd(run_path):
-    starting_directory = os.getcwd()
-    os.chdir(run_path)
-    yield
-    os.chdir(starting_directory)
-
-
-def find_unsmry(basepath: Path) -> Path:
-    def _is_unsmry(base: str, path: str) -> bool:
-        if "." not in path:
-            return False
-        splitted = path.split(".")
-        return splitted[-2].endswith(base) and splitted[-1].lower() in [
-            "unsmry",
-            "funsmry",
-        ]
-
-    dir = basepath.parent
-    base = basepath.name
-    candidates: List[str] = list(
-        filter(lambda x: _is_unsmry(base, x), os.listdir(dir or "."))
-    )
-    if not candidates:
-        raise ValueError(f"Could not find any unsmry matching case path {basepath}")
-    if len(candidates) > 1:
-        raise ValueError(
-            f"Ambiguous reference to unsmry in {basepath}, could be any of {candidates}"
-        )
-    return Path(dir) / candidates[0]
-
-
-def await_completed_unsmry_file(
-    smry_path: Path, max_wait: float = 15, poll_interval: float = 1.0
-) -> float:
-    """This function will wait until the provided smry file does not grow in size
-    during one poll interval.
-
-    Such a wait is sometimes needed when different MPI hosts write data to a shared
-    disk system.
-
-    If the file does not exist or is completely unreadable to resfo, this function
-    will timeout to max_wait. If NOSIM is included, this will happen.
-
-    Size is defined in terms of readable data elementes through resfo.
-
-    This function will always wait for at least one poll interval, the polling
-    interval is specified in seconds.
-
-    The return value is the waited time (in seconds)"""
-    start_time = datetime.datetime.now()
-    prev_len = 0
-    while (datetime.datetime.now() - start_time).total_seconds() < max_wait:
-        try:
-            resfo_sum = [r.read_keyword() for r in resfo.lazy_read(smry_path)]
-        except Exception:
-            time.sleep(poll_interval)
-            continue
-
-        current_len = len(resfo_sum)
-        if prev_len == current_len:
-            # smry file is regarded complete
-            break
-        else:
-            prev_len = max(prev_len, current_len)
-
-        time.sleep(poll_interval)
-
-    return (datetime.datetime.now() - start_time).total_seconds()
-
-
-class EclRun:
-    """Wrapper class to run Eclipse simulations.
-
-    The EclRun class is a small wrapper class which is used to run Eclipse
-    simulations. It will load a configuration, i.e. where the binary is
-    installed and so on, from an instance of the EclConfig class.
-
-    The main method is the runEclipse() method which will:
-
-      4. Run the system provided eclrun binary
-      5. Parse the output .PRT / .ECLEND file to check for errors.
-
-    If the simulation fails the runEclipse() method will raise an exception.
-
-    The class is called EclRun, and the main focus has been on running Eclipse
-    simulations, but it should also handle "eclipse-like" simulators, e.g. the
-    simulator OPM/flow.
-
-    To actually create an executable script based on this class could in it's
-    simplest form be:
-
-       #!/usr/bin/env python
-       import sys
-       from .ecl_run import EclRun
-
-       run = EclRun()
-       run.runEclipse( )
-
-
-    """
-
-    def __init__(
-        self,
-        ecl_case: str,
-        num_cpu: int = 1,
-        check_status: bool = True,
-        summary_conversion: bool = False,
-    ):
-        pm = plugin_manager.ErtPluginManager()
-        self.eclrun_abspath = shutil.which(
-            "eclrun",
-            path=os.pathsep.join(
-                pm.get_forward_model_paths() + os.getenv("PATH", "").split(os.pathsep)
-            ),
-        )
-        if self.eclrun_abspath is None:
-            raise RuntimeError("eclrun not installed")
-
-        self.check_status = check_status
-        self.num_cpu = int(num_cpu)
-        self.summary_conversion = summary_conversion
-
-        # Dechipher the ecl_case argument.
-        input_arg = ecl_case
-        (_, ext) = os.path.splitext(input_arg)
-        if ext and ext in [".data", ".DATA"]:
-            data_file = input_arg
-        elif input_arg.islower():
-            data_file = input_arg + ".data"
-        else:
-            data_file = input_arg + ".DATA"
-
-        if not os.path.isfile(data_file):
-            raise IOError(f"No such file: {data_file}")
-
-        (self.run_path, self.data_file) = os.path.split(data_file)
-        (self.base_name, ext) = os.path.splitext(self.data_file)
-
-        if self.run_path is None:
-            self.run_path = os.getcwd()
-        else:
-            self.run_path = os.path.abspath(self.run_path)
-
-    def runPath(self):
-        return self.run_path
-
-    def baseName(self):
-        return self.base_name
-
-    @property
-    def prt_path(self):
-        return Path(self.run_path) / (self.baseName() + ".PRT")
-
-    def numCpu(self):
-        return self.num_cpu
-
-    def initMPI(self):
-        # If the environment variable LSB_MCPU_HOSTS is set we assume the job is
-        # running on LSF - otherwise we assume it is running on the current host.
-        #
-        # If the LSB_MCPU_HOSTS variable is indeed set it will be a string like this:
-        #
-        #       host1 num_cpu1 host2 num_cpu2 ...
-        #
-        # i.e. an alternating list of hostname & number of
-        # cpu. Alternatively/in addition the environment variable
-        # LSB_HOSTS can be used. This variable is simply:
-        #
-        #       host1  host1  host2  host3
-
-        LSB_MCPU_HOSTS = os.getenv("LSB_MCPU_HOSTS")
-        LSB_HOSTS = os.getenv("LSB_HOSTS")
-
-        if LSB_MCPU_HOSTS or LSB_HOSTS:
-            LSB_MCPU_machine_list = make_LSB_MCPU_machine_list(LSB_MCPU_HOSTS)
-            LSB_machine_list = make_LSB_machine_list(LSB_HOSTS)
-
-            if len(LSB_MCPU_machine_list) == self.num_cpu:
-                machine_list = LSB_MCPU_machine_list
-            elif len(LSB_machine_list) == self.num_cpu:
-                machine_list = LSB_machine_list
-            else:
-                raise EclError(
-                    "LSF / MPI problems. "
-                    f"Asked for:{self.num_cpu} cpu. "
-                    f'LSB_MCPU_HOSTS: "{LSB_MCPU_HOSTS}"  LSB_HOSTS: "{LSB_HOSTS}"'
-                )
-        elif os.getenv("SLURM_JOB_NODELIST"):
-            machine_list = make_SLURM_machine_list(
-                os.getenv("SLURM_JOB_NODELIST"), os.getenv("SLURM_TASKS_PER_NODE")
-            )
-            if len(machine_list) != self.num_cpu:
-                raise EclError(
-                    f"SLURM / MPI problems - asked for {self.num_cpu} - "
-                    f"got {len(machine_list)} nodes"
-                )
-        else:
-            localhost = socket.gethostname()
-            machine_list = [localhost] * self.num_cpu
-
-        self.machine_file = f"{self.base_name}.mpi"
-        with open(self.machine_file, "w", encoding="utf-8") as filehandle:
-            for host in machine_list:
-                filehandle.write(f"{host}\n")
-
-    def _get_run_command(self):
-        assert self.eclrun_abspath is not None
-        summary_conversion = "yes" if self.summary_conversion else "no"
-        return [
-            self.eclrun_abspath,
-            "-v",
-            eclrun_config.version,
-            eclrun_config.simulator_name,
-            f"{self.base_name}.DATA",
-            "--summary-conversion",
-            summary_conversion,
-        ]
-
-    def _run_eclrun_binary(self) -> int:
-        with pushd(self.run_path):
-            if not os.path.exists(self.data_file):
-                raise IOError(f"Can not find data_file:{self.data_file}")
-            if not os.access(self.data_file, os.R_OK):
-                raise OSError(f"Can not read data file:{self.data_file}")
-
-            return subprocess.run(
-                self._get_run_command(),
-                check=False,
-            ).returncode
-
-    LICENSE_FAILURE_RETRY_INITIAL_SLEEP = 90
-    LICENSE_RETRY_STAGGER_FACTOR = 60
-    LICENSE_RETRY_BACKOFF_EXPONENT = 3
-
-    def runEclipse(self, retries_left=3, backoff_sleep=None):
-        backoff_sleep = (
-            self.LICENSE_FAILURE_RETRY_INITIAL_SLEEP
-            if backoff_sleep is None
-            else backoff_sleep
-        )
-        return_code = self._run_eclrun_binary()
-
-        OK_file = os.path.join(self.run_path, f"{self.base_name}.OK")
-        if not self.check_status:
-            with open(OK_file, "w", encoding="utf-8") as f:
-                f.write("ECLIPSE simulation complete - NOT checked for errors.")
-        else:
-            if return_code != 0:
-                command = self._get_run_command()
-                raise subprocess.CalledProcessError(return_code, command)
-
-            try:
-                self.assertECLEND()
-            except EclError as err:
-                if err.failed_due_to_license_problems() and retries_left > 0:
-                    time_to_wait = backoff_sleep + int(
-                        random() * self.LICENSE_RETRY_STAGGER_FACTOR
-                    )
-                    sys.stderr.write(
-                        "ECLIPSE failed due to license failure "
-                        f"retrying in {time_to_wait} seconds\n"
-                    )
-                    time.sleep(time_to_wait)
-                    self.runEclipse(
-                        retries_left=retries_left - 1,
-                        backoff_sleep=int(
-                            backoff_sleep * self.LICENSE_RETRY_BACKOFF_EXPONENT
-                        ),
-                    )
-                    return
-                else:
-                    raise err from None
-            if self.num_cpu > 1:
-                await_completed_unsmry_file(
-                    find_unsmry(Path(self.run_path) / self.base_name)
-                )
-
-            with open(OK_file, "w", encoding="utf-8") as f:
-                f.write("ECLIPSE simulation OK")
-
-    def assertECLEND(self):
-        tail_length = 5000
-        result = self.readECLEND()
-        if result.errors > 0:
-            error_list = self.parseErrors()
-            sep = "\n\n...\n\n"
-            error_and_slave_msg = sep.join(error_list)
-            extra_message = ""
-            error_messages = [
-                error for error in error_list if not "STARTING SLAVE" in str(error)
-            ]
-            if result.errors != len(error_messages):
-                extra_message = (
-                    f"\n\nWarning, mismatch between stated Error count ({result.errors}) "
-                    f"and number of ERROR messages found in PRT ({len(error_messages)})."
-                    f"\n\nTail ({tail_length} bytes) of PRT-file {self.prt_path}:\n\n"
-                ) + tail_textfile(self.prt_path, 5000)
-
-            raise EclError(
-                "Eclipse simulation failed with:"
-                f"{result.errors:d} errors:\n\n{error_and_slave_msg}{extra_message}"
-            )
-
-        if result.bugs > 0:
-            raise EclError(f"Eclipse simulation failed with:{result.bugs:d} bugs")
-
-    def readECLEND(self):
-        error_regexp = re.compile(r"^\s*Errors\s+(\d+)\s*$")
-        bug_regexp = re.compile(r"^\s*Bugs\s+(\d+)\s*$")
-
-        report_file = os.path.join(self.run_path, f"{self.base_name}.ECLEND")
-        if not os.path.isfile(report_file):
-            report_file = self.prt_path
-
-        errors = None
-        bugs = None
-        with open(report_file, "r", encoding="utf-8") as filehandle:
-            for line in filehandle.readlines():
-                error_match = re.match(error_regexp, line)
-                if error_match:
-                    errors = int(error_match.group(1))
-
-                bug_match = re.match(bug_regexp, line)
-                if bug_match:
-                    bugs = int(bug_match.group(1))
-        if errors is None:
-            raise ValueError(f"Could not read errors from {report_file}")
-        if bugs is None:
-            raise ValueError(f"Could not read bugs from {report_file}")
-
-        return EclipseResult(errors=errors, bugs=bugs)
-
-    def parseErrors(self) -> List[str]:
-        """Extract multiline ERROR messages from the PRT file"""
-        error_list = []
-        error_e100_regexp = re.compile(error_pattern_e100, re.MULTILINE)
-        error_e300_regexp = re.compile(error_pattern_e300, re.MULTILINE)
-        slave_started_regexp = re.compile(slave_started_pattern, re.MULTILINE)
-        with open(self.prt_path, "r", encoding="utf-8") as filehandle:
-            content = filehandle.read()
-
-        for regexp in [error_e100_regexp, error_e300_regexp, slave_started_regexp]:
-            offset = 0
-            while True:
-                match = regexp.search(content[offset:])
-                if match:
-                    error_list.append(
-                        content[offset + match.start() : offset + match.end()]
-                    )
-                    offset += match.end()
-                else:
-                    break
-
-        return error_list
-
-
-def run(config: EclConfig, argv):
-    parser = ArgumentParser()
-    parser.add_argument("ecl_case")
-    parser.add_argument("-v", "--version", dest="version", type=str)
-    parser.add_argument("-n", "--num-cpu", dest="num_cpu", type=int, default=1)
-    parser.add_argument(
-        "-i", "--ignore-errors", dest="ignore_errors", action="store_true"
-    )
-    parser.add_argument(
-        "--summary-conversion", dest="summary_conversion", action="store_true"
-    )
-
-    options = parser.parse_args(argv)
-
-    try:
-        run = EclRun(
-            options.ecl_case,
-            num_cpu=options.num_cpu,
-            check_status=not options.ignore_errors,
-            summary_conversion=options.summary_conversion,
-        )
-        run.runEclipse()
-    except EclError as msg:
-        print(msg, file=sys.stderr)
-        sys.exit(-1)
-
-
-def tail_textfile(file_path: Path, num_chars: int) -> str:
-    if not file_path.exists():
-        return f"No output file {file_path}"
-    with open(file_path, encoding="utf-8") as file:
-        file.seek(0, 2)
-        file_end_position = file.tell()
-        seek_position = max(0, file_end_position - num_chars)
-        file.seek(seek_position)
-        return file.read()[-num_chars:]
