+++ test-data/ert/flow_example/flow.ert
@@ -16,4 +16,4 @@ GEN_KW FIELD_PROPERTIES field_properties_priors
 
 FORWARD_MODEL TEMPLATE_RENDER(<INPUT_FILES>=parameters.json, <TEMPLATE_FILE>=<CONFIG_PATH>/resources/SPE1.DATA.jinja2, <OUTPUT_FILE>=SPE1.DATA)
 
-FORWARD_MODEL FLOW
+FORWARD_MODEL ECLIPSE100(<VERSION>=2019.3)
